Update RetexturedHelper#addTagVariants for how I ended up filling tabs
The function no longer adds the no NBT item, instead returning a boolean indicating if items were added that can be used to add the default if desired.
Additionally, it takes two tab visibilities, one for the first item and one for consecutive items. Second can be set to null to stop variants after the first from showing.