Remove IGenericLoader in favor of using RecordLoadable directly
As part of this change, GenericLoaderRegistry now implements RecordLoadable, though its not really viable to use it as an entry in a generic loader registry unless you were to change the type key.
This notably adds some nice utilities to the loader registry from record loadable and just simplifies a lot of things. You can still directly create a RecordLoadable if you prefer the old style instead of the builders.