Propagate context through nested loadables
Before it was exclusive to record loadables. Now all of the API supports it, passing it on through each step
Helper methods were created that automatically set the context to TypedMap.EMPTY to allow old usages of loadables outside the API to work.
As a result of this, removed ContextSreamable, its no longer needed