Improve tag conditions
Added TagFilledCondition, inverted from TagEmptyCondition. 99% of the time we want to check not empty over empty
Both TagFilledCondition and TagEmptyCondition are now registered into loot tables
Added TagCombination, handling both difference and intersection queries. Note unlike the old tinkers version, you need to manually combine tags to use them in an OR on the difference.
All tag conditions registry defaults to item if unset, as item is the one we use these for the most