Fix oddly named JSON field for fluid container transfer
The result of emptying the container was a stack named "filled", which is misleading as to its function.
For the filling variant, filled was better, but it was still inconsistent as the start was "input".
To avoid breaking old datapacks (e.g. tinkers), the old field name is still supported as deprecated.