Fix tooltips going out of bounds on right side pages
Issue is we offset the tooltip renderer for the right side, causing it to miscalculate the true screen bounds.
This is caused by tooltip rendering using the global window size instead of any size we can set.
Fix by reimplementing the text wrapping and tooltip positioning relative to our book page size