Fix content grouping transformer mishandling padding
Padding got added to the end of the section in the fallback case, which should never happen but just in case best to do it right.
Was also do a looped index add, which is less efficient than doing an indexed add with a list in most cases; saves shifting down every element N times.